app/requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
pytest==7.4.3
requests==2.31.0